+++ lib/vlc_player.dart
@@ -1,4 +1,5 @@
 export 'src/vlc_player_error.dart';
+export 'src/vlc_media_source.dart';
 export 'src/vlc_media_info.dart';
 export 'src/vlc_player.dart';
 export 'src/vlc_player_controller.dart';